Clean up MetadataCollection and builder
There's a lot of extra code for building MetadataCollection that really isn't used much. This change removes a lot of that and unifies code that needs to build MetadataCollections on MetadataBuilder.

There are a few tests that are now unneeded after this change because the `Metadata` property has been removed from all tag helper builder types. Now, everything goes through `SetMetadata(...)` on tag helper builders.