IDocumentSnapshot: Add GetSourceAsync and TryGetSource
This change adds GetSourceAsync and TryGetSource methods to IDocumentSnapshot that return a RazorSourceDocument. The GetTextAsync and TryGetText methods have been converted to extension methods on IDocumentSnapshot.

This change is fairly mechanical but results in a lot of test changes because there are so many IDocumentSnapshot mocks that have been copied-and-pasted across tests.