DocumentState: Improve text loading and fix IDocumentSnapshot forking
When a DocumentState is updated with a new SourceText, we can avoid wrapping the SourceText in a TextLoader if its text hasn't been loaded yet.

In addition, this fixes the DocumentSnapshot.WithText(...) method to correctly fork its parent ProjectSnapshot. Previously, after calling IDocumentSnapshot, calling IDocumentSnapshot.Project.TryGetDocument(...) for the same document would return the older document. Now it correctly returns the forked version.

The TestDocumentSnapshot helper has been updated to correctly fork the snapshot and be parented by a real project snapshot.