Fix up tests using bad file paths
Many tests were creating URIs with bad paths that fail with the TestDocumentSnapshot changes in the previous commit.

These bad URIs usually take the form of "file://path/test.razor". What's wrong about that, you ask? Glad you asked!

When we process URIs created with paths like "file://path/test.razor" on Windows, we end up getting a local path like "\\path\test.razor". This is because the path isn't rooted on Windows, so it's assumed to be a UNC path. HOWEVER, this is an invalid UNC file path because there should be at least _three_ parts:

\\<server-name>\<share-name>\<file-path>

Instead, we just have two parts, which isn't a legal path. This causes tests to fail with the changes to TestDocumentSnapshot. Essentially, when ProjectState attempts to create a ProjectEngine, it calls Path.GetDirectoryName(...) on the project file path. However, because it's an invalid UNC path, Path.GetDirectoryName(...) returns null.

This change updates each instance of an invalid URI file path to include an extra part, allowing tests to pass.