Merge RazorProjectEngineBuilderExtensions classes in compiler
There are two RazorProjectEngineBuilderExtensions classes in the compiler in different namespaces. This merges the two, which may be a breaking change for the RazorSDK, which uses the SetCSharpLanguageVersion extension method.

This change also merges the RazorProjectEngineBuilderExtensionsTest files and tweaks the SetCSharpLanguageVersion test to be a bit more useful.

Finally, a few using directives have been removed since SetCSharpLanguageVersion now resolves from a different namespace.